[BUGFIX] Fixed class config for SliderNews
The class configuration for the Extbase persistence layer was simplified
to reduce side effects.

Since we do not need special subclasses in our list we always use the
slider news class for all news records. For this to work all other
class settings are removed and only the SliderNews class is configured.

The obsolete NewsOther class was removed and the displayNews property
was moved to the SliderNews class. The class type of the displayNews
property was also changed to SliderNews so that referenced news are
displayed correctly in the slider.

Some obsolete closing PHP tags have been removed.

Relates: #5669