[BUGFIX] Fix news detail rendering
After the class configuration was cleaned up the rich teaser was not
available in the detail view any more because the default news subclass
is not overwritten with the richteaser news domain model.

To make the rich teaser work again in the detail view the detail action
is overwritten in the news_slideit news controller. The action now
expects a richteaser news domain model and passes it on to the parent
detail action.

Note! This currently does not work with the SliderNews domain object
because of a mapping error for the sliderImageFromContent property.

Relates: #5669