[TASK] News limit only for simple list
The limit setting for news is now only available in the simple list.

Resolved by using the displayCond configuration in the overwritten
FlexForm of the news Extension. The <settings.limit> field had to be
moved to the main sheet because in the field display conditions only
field values of the same sheet are accessible, see also:
https://forge.typo3.org/issues/42333

Relates: #5669
Change-Id: I8ba9022a22401eab7ff93c5decb6feea98760a5f