[FEATURE] Implemented RSS feeds for news
The RSS feed for news are implemented. With this change every list,
simple list or slider plugin gets a link to an RSS feed with a matching
title.

The feed settings can be modified by TypoScript and overwritten with
new FlexForm settings (to make localization possible).

To be able to generate unique IDs for news the original news UID will
be set to the originalUid property in the referenced news.

A new page type is introduced for rendering the RSS feed. If this page
type is active the content element rendering will be simplified and all
URLs will be made absolute.

Additionally the news_slideit Extension code was cleaned up and improved.

Relates: #5669, #5668, #6820
Change-Id: I3c2ab3a1f03ab04329ccf9ba88634c880eec7810